Fix the multiaddr reexport
